+++ semcore/wizard/src/index.js
@@ -1,2 +1,0 @@
-export { default } from './Wizard';
-export * from './Wizard';