+++ stories/components/base-trigger/advanced/base-trigger.stories.tsx
@@ -1,0 +1,33 @@
+import type { Meta, StoryObj } from '@storybook/react-vite';
+
+import BaseTriggerEllipsisExample from './examples/base-trigger-ellipsis';
+import ButtonTriggerEllipsisExample from './examples/button-trigger-ellipsis';
+import FilterTriggerEllipsisExample from './examples/filter-trigger-ellipsis';
+import FilterTriggerWithCounterExample from './examples/filter-trigger_with_counter';
+import LinkTriggerAndEllipsisExample from './examples/link-trigger-ellipsis';
+
+const meta: Meta = {
+  title: 'Components/Base Trigger/Advanced',
+};
+
+export const LinkTriggerAndEllipsis: StoryObj = {
+  render: LinkTriggerAndEllipsisExample,
+};
+
+export const BaseTriggerEllipsis: StoryObj = {
+  render: BaseTriggerEllipsisExample,
+};
+
+export const ButtonTriggerEllipsis: StoryObj = {
+  render: ButtonTriggerEllipsisExample,
+};
+
+export const FilterTriggerEllipsis: StoryObj = {
+  render: FilterTriggerEllipsisExample,
+};
+
+export const FilterTriggerWithCounter: StoryObj = {
+  render: FilterTriggerWithCounterExample,
+};
+
+export default meta;